hello

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 """ Moudule """
-from typing import Callable, Iterator, Union, Optional, List
+from typing import  Union, List
 
 
-def sum_mixed_list(mxd_lst: List[Union[float,int]]) -> float:
+def sum_mixed_list(mxd_lst: List[Union[int,float]]) -> float:
     """ Function """
     sum = 0
